gcc 3.1 needs a static_cast<int> here as the != operator is ambiguous
svn path=/trunk/koffice/; revision=159213

+++ kpresenter/kptextobject.cc
@@ -1027,7 +1027,7 @@ void KPTextObject::saveParagLayout( const KoParagLayout& layout, QDomElement & p
         element.setAttribute( "width", (*it).ptWidth );
     }
 
-    if(layout.shadowDistance!=0 || layout.shadowDirection!=KoParagLayout::SD_RIGHT_BOTTOM)
+    if(layout.shadowDistance!=0 || layout.shadowDirection!=static_cast<int>(KoParagLayout::SD_RIGHT_BOTTOM))
     {
         element = doc.createElement( "SHADOW" );
         parentElem.appendChild( element );
